Moved useMountSignal to render
