+++ src/main/index.ts
@@ -8,9 +8,8 @@ export * from './memo';
 export * from './metronome';
 export * from './polling';
 export * from './ref';
-export * from './rerender';
+export * from './render';
 export * from './time';
 export * from './areHookInputsEqual';
-export * from './useMountSignal';
 export * from './useToggle';
 export * from './shared-types';
