+++ src/main/render/index.ts
@@ -1,2 +1,3 @@
+export * from './useMountSignal';
 export * from './useRerender';
 export * from './useRerenderSchedule';
