+++ src/test/render/useMountSignal.test.ts
@@ -1,5 +1,5 @@
 import {renderHook} from '@testing-library/react-hooks/native';
-import {useMountSignal} from '../main';
+import {useMountSignal} from '../../main';
 
 describe('useMountSignal', () => {
